File reorganization and checking in test recordings after last test pass.
