+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/tests/latest/recordings/test_prepare_publish_should_raise_cli_error_when_version_is_v4.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
-    body: '{"location": "westus", "tags": {"product": "azurecli", "cause": "automation",
-      "date": "2018-11-27T04:30:23Z"}}'
+    body: '{"tags": {"date": "2018-11-27T06:51:58Z", "cause": "automation", "product":
+      "azurecli"}, "location": "westus"}'
     headers:
       Accept: [application/json]
       Accept-Encoding: ['gzip, deflate']
@@ -10,23 +10,23 @@ interactions:
       Content-Length: ['110']
       Content-Type: [application/json; charset=utf-8]
       ParameterSetName: [--location --name --tag]
-      User-Agent: [python/3.7.0 (Windows-10-10.0.17763-SP0) msrest/0.6.1 msrest_azure/0.4.34
+      User-Agent: [python/3.5.2 (Windows-10-10.0.15063-SP0) msrest/0.6.1 msrest_azure/0.4.34
           resourcemanagementclient/2.0.0 Azure-SDK-For-Python AZURECLI/2.0.51]
       accept-language: [en-US]
     method: PUT
     uri: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/clitest.rg000001?api-version=2018-05-01
   response:
-    body: {string: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/clitest.rg000001","name":"clitest.rg000001","location":"westus","tags":{"product":"azurecli","cause":"automation","date":"2018-11-27T04:30:23Z"},"properties":{"provisioningState":"Succeeded"}}'}
+    body: {string: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/clitest.rg000001","name":"clitest.rg000001","location":"westus","tags":{"date":"2018-11-27T06:51:58Z","cause":"automation","product":"azurecli"},"properties":{"provisioningState":"Succeeded"}}'}
     headers:
       cache-control: [no-cache]
       content-length: ['274']
       content-type: [application/json; charset=utf-8]
-      date: ['Tue, 27 Nov 2018 04:30:26 GMT']
+      date: ['Tue, 27 Nov 2018 06:51:58 GMT']
       expires: ['-1']
       pragma: [no-cache]
       strict-transport-security: [max-age=31536000; includeSubDomains]
       x-content-type-options: [nosniff]
-      x-ms-ratelimit-remaining-subscription-writes: ['1199']
+      x-ms-ratelimit-remaining-subscription-writes: ['1197']
     status: {code: 201, message: Created}
 - request:
     body: null
@@ -37,7 +37,7 @@ interactions:
       Connection: [keep-alive]
       Content-Length: ['0']
       ParameterSetName: [-g -n]
-      User-Agent: [python/3.7.0 (Windows-10-10.0.17763-SP0) msrest/0.6.1 msrest_azure/0.4.34
+      User-Agent: [python/3.5.2 (Windows-10-10.0.15063-SP0) msrest/0.6.1 msrest_azure/0.4.34
           azure-mgmt-botservice/0.1.0 Azure-SDK-For-Python AZURECLI/2.0.51]
       accept-language: [en-US]
     method: DELETE
@@ -46,12 +46,12 @@ interactions:
     body: {string: ''}
     headers:
       cache-control: [no-cache]
-      date: ['Tue, 27 Nov 2018 04:30:25 GMT']
+      date: ['Tue, 27 Nov 2018 06:51:59 GMT']
       expires: ['-1']
       pragma: [no-cache]
       strict-transport-security: [max-age=31536000; includeSubDomains]
       x-content-type-options: [nosniff]
-      x-ms-ratelimit-remaining-subscription-deletes: ['14999']
+      x-ms-ratelimit-remaining-subscription-deletes: ['14998']
     status: {code: 204, message: No Content}
 - request:
     body: null
@@ -62,7 +62,7 @@ interactions:
       Connection: [keep-alive]
       Content-Length: ['0']
       ParameterSetName: [-g -n]
-      User-Agent: [python/3.7.0 (Windows-10-10.0.17763-SP0) msrest/0.6.1 msrest_azure/0.4.34
+      User-Agent: [python/3.5.2 (Windows-10-10.0.15063-SP0) msrest/0.6.1 msrest_azure/0.4.34
           azure-mgmt-botservice/0.1.0 Azure-SDK-For-Python AZURECLI/2.0.51]
       accept-language: [en-US]
     method: DELETE
@@ -71,12 +71,12 @@ interactions:
     body: {string: ''}
     headers:
       cache-control: [no-cache]
-      date: ['Tue, 27 Nov 2018 04:30:26 GMT']
+      date: ['Tue, 27 Nov 2018 06:51:59 GMT']
       expires: ['-1']
       pragma: [no-cache]
       strict-transport-security: [max-age=31536000; includeSubDomains]
       x-content-type-options: [nosniff]
-      x-ms-ratelimit-remaining-subscription-deletes: ['14999']
+      x-ms-ratelimit-remaining-subscription-deletes: ['14997']
     status: {code: 204, message: No Content}
 - request:
     body: null
@@ -88,7 +88,7 @@ interactions:
       Content-Length: ['0']
       Content-Type: [application/json; charset=utf-8]
       ParameterSetName: [--name --yes --no-wait]
-      User-Agent: [python/3.7.0 (Windows-10-10.0.17763-SP0) msrest/0.6.1 msrest_azure/0.4.34
+      User-Agent: [python/3.5.2 (Windows-10-10.0.15063-SP0) msrest/0.6.1 msrest_azure/0.4.34
           resourcemanagementclient/2.0.0 Azure-SDK-For-Python AZURECLI/2.0.51]
       accept-language: [en-US]
     method: DELETE
@@ -98,12 +98,12 @@ interactions:
     headers:
       cache-control: [no-cache]
       content-length: ['0']
-      date: ['Tue, 27 Nov 2018 04:30:32 GMT']
+      date: ['Tue, 27 Nov 2018 06:52:01 GMT']
       expires: ['-1']
-      location: ['https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/operationresults/eyJqb2JJZCI6IlJFU09VUkNFR1JPVVBERUxFVElPTkpPQi1DTElURVNUOjJFUkdKSFAzSVpBNFhWLVdFU1RVUyIsImpvYkxvY2F0aW9uIjoid2VzdHVzIn0?api-version=2018-05-01']
+      location: ['https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/operationresults/eyJqb2JJZCI6IlJFU09VUkNFR1JPVVBERUxFVElPTkpPQi1DTElURVNUOjJFUkdCSUJWTERGVDQzLVdFU1RVUyIsImpvYkxvY2F0aW9uIjoid2VzdHVzIn0?api-version=2018-05-01']
       pragma: [no-cache]
       strict-transport-security: [max-age=31536000; includeSubDomains]
       x-content-type-options: [nosniff]
-      x-ms-ratelimit-remaining-subscription-deletes: ['14999']
+      x-ms-ratelimit-remaining-subscription-deletes: ['14997']
     status: {code: 202, message: Accepted}
 version: 1
